- 使用官方插件模板重构项目结构 - 增加国际化支持 - 修复action功能开关失效和一些别的bug
