+++ CHANGELOG.md
@@ -1,0 +1,53 @@
+<!-- Keep a Changelog guide -> https://keepachangelog.com -->
+
+# Activate Power Mode Changelog
+
+## [0.1.9] - 2026-03-06
+
+- 替换废弃的api
+- 修复action功能开关失效的问题
+
+## [0.1.8]
+
+- 优化 Combo 开关关闭 Combo 计算显示
+
+## [0.1.7]
+
+- 修复 Combo 开关关闭 Combo 计算还是显示问题
+- 粒子颜色根据 EditorHighlighter 获取颜色显示
+
+## [0.1.6]
+
+- 增加最大 Combo 数显示
+- 增加 Combo 连击数显示
+- 增加 Combo 时间间隔显示
+
+## [0.1.5]
+
+- 修复删除没有粒子效果
+- 修复快速敲击时存在残影残留
+
+## [0.1.4]
+
+- 增加 setting 选择窗口
+- 增加粒子颜色配置项
+- 整理 action 的重复代码
+
+## [0.1.3]
+
+- 修复在点击 enable 的时候抛出 java.util.ConcurrentModificationException
+- 修复在 ActivatePowerModeManage destroy 的时候 抛出 java.lang.IllegalStateException
+- 继续优化多 Project 的情况
+
+## [0.1.2]
+
+- 增加一个 Combo 值，达到一定 Combo 值自动开启效果
+- 优化多个 Project 和多个 Editor 的情况实现
+
+## [Unreleased]
+
+### Planned Features
+
+- 把 Combo, MaxCombo 可视化出来
+- 做设置数据持久化
+- 增加设置参数界面
