+++ gradle.properties
@@ -1,0 +1,33 @@
+# IntelliJ Platform Artifacts Repositories -> https://plugins.jetbrains.com/docs/intellij/intellij-artifacts.html
+
+pluginGroup = org.jetbrains.plugins.template
+pluginName = activate-power-mode
+pluginRepositoryUrl = https://github.com/JetBrains/intellij-platform-plugin-template
+# SemVer format -> https://semver.org
+pluginVersion = 0.1.9
+
+# Supported build number ranges and IntelliJ Platform versions -> https://plugins.jetbrains.com/docs/intellij/build-number-ranges.html
+pluginSinceBuild = 253
+
+# IntelliJ Platform Properties -> https://plugins.jetbrains.com/docs/intellij/tools-gradle-intellij-plugin.html#configuration-intellij-extension
+platformVersion = 2025.3.3
+
+# Plugin Dependencies -> https://plugins.jetbrains.com/docs/intellij/plugin-dependencies.html
+# Example: platformPlugins = com.jetbrains.php:203.4449.22, org.intellij.scala:2023.3.27@EAP
+platformPlugins =
+# Example: platformBundledPlugins = com.intellij.java
+platformBundledPlugins =
+# Example: platformBundledModules = intellij.spellchecker
+platformBundledModules =
+
+# Gradle Releases -> https://github.com/gradle/gradle/releases
+gradleVersion = 9.3.1
+
+# Opt-out flag for bundling Kotlin standard library -> https://jb.gg/intellij-platform-kotlin-stdlib
+kotlin.stdlib.default.dependency = false
+
+# Enable Gradle Configuration Cache -> https://docs.gradle.org/current/userguide/configuration_cache.html
+org.gradle.configuration-cache = true
+
+# Enable Gradle Build Cache -> https://docs.gradle.org/current/userguide/build_cache.html
+org.gradle.caching = true
