--- resources/META-INF/plugin.xml
@@ -1,129 +1,0 @@
-<idea-plugin>
-    <id>com.jiyuanime.ActivatePowerModeApplicationPlugin</id>
-    <name>activate-power-mode</name>
-    <version>0.1.8</version>
-    <vendor email="" url="https://github.com/ViceFantasyPlace/activate-power-mode">
-        ViceFantasyPlace
-    </vendor>
-
-    <description>
-        <![CDATA[
-          <h3>activate-power-mode for IDEA.</h3>
-          <ul>
-			<li>根据Atom的插件activate-power-mode的效果移植到IDEA上</li>
-		  </ul>
-        ]]>
-    </description>
-
-    <change-notes>
-        <![CDATA[
-          <h3>v0.1.8</h3>
-          <ul>
-            <li>优化Combo开关关闭Combo计算显示</li>
-          </ul>
-
-          <h3>v0.1.7</h3>
-          <ul>
-            <li>修复Combo开关关闭Combo计算还是显示问题</li>
-            <li>粒子颜色根据EditorHighlighter获取颜色显示</li>
-          </ul>
-
-          <h3>v0.1.6</h3>
-          <ul>
-            <li>增加最大Combo数显示</li>
-            <li>增加Combo连击数显示</li>
-            <li>增加Combo时间间隔显示</li>
-          </ul>
-
-          <h3>v0.1.5</h3>
-          <ul>
-            <li>修复删除没有粒子效果</li>
-            <li>修复快速敲击时存在残影残留</li>
-          </ul>
-
-          <h3>v0.1.4</h3>
-          <ul>
-            <li>增加setting选择窗口</li>
-            <li>增加粒子颜色配置项</li>
-            <li>整理action的重复代码</li>
-          </ul>
-
-          <h3>v0.1.3</h3>
-          <ul>
-            <li>修复在点击enable的时候抛出 java.util.ConcurrentModificationException</li>
-            <li>修复在ActivatePowerModeManage destroy 的时候 抛出 java.lang.IllegalStateException</li>
-            <li>继续优化多Project的情况</li>
-          </ul>
-
-          <h3>v0.1.2</h3>
-          <ul>
-            <li>增加一个Combo值，达到一定Combo值自动开启效果</li>
-            <li>优化多个Project和多个Editor的情况实现</li>
-          </ul>
-
-          <h3>Features.</h3>
-          <ul
-            <li>把Combo, MaxCombo 可视化出来</li>
-            <li>做设置数据持久化</li>
-            <li>增加设置参数界面</li>
-          </ul>
-        ]]>
-    </change-notes>
-
-    <!-- please see http://www.jetbrains.org/intellij/sdk/docs/basics/getting_started/build_number_ranges.html for description -->
-    <idea-version since-build="141.0"/>
-
-    <!-- please see http://www.jetbrains.org/intellij/sdk/docs/basics/getting_started/plugin_compatibility.html
-         on how to target different products -->
-    <!-- uncomment to enable plugin in all products
-    <depends>com.intellij.modules.lang</depends>
-    -->
-    <depends>com.intellij.modules.lang</depends>
-
-    <extensions defaultExtensionNs="com.intellij">
-        <applicationService serviceInterface="com.jiyuanime.config.Config"
-                            serviceImplementation="com.jiyuanime.config.Config"/>
-        <applicationConfigurable instance="com.jiyuanime.ui.Setting"/>
-    </extensions>
-
-    <actions>
-        <!-- Add your actions here -->
-        <group id="ActivatePowerModeGroup" popup="true"
-               text="activate-power-mode">
-            <add-to-group group-id="WindowMenu" anchor="first"/>
-
-            <action id="ActivatePowerEnable" class="com.jiyuanime.action.ActivatePowerEnableAction"
-                    text="enable"
-                    description="enable">
-                <keyboard-shortcut keymap="$default" first-keystroke="shift ctrl alt A"/>
-            </action>
-
-            <action id="ShakeSwitch" class="com.jiyuanime.action.ShakeSwitchAction" text="shake"
-                    description="shake">
-                <keyboard-shortcut keymap="$default" first-keystroke="shift ctrl alt S"/>
-            </action>
-
-            <action id="ParticleSwitch" class="com.jiyuanime.action.ParticleSwitchAction"
-                    text="particle"
-                    description="particle">
-                <keyboard-shortcut keymap="$default" first-keystroke="shift ctrl alt P"/>
-            </action>
-
-            <action id="ComboSwitchAction" class="com.jiyuanime.action.ComboSwitchAction"
-                    text="combo"
-                    description="combo"/>
-
-            <action id="ColorfulSwitch" class="com.jiyuanime.action.ColorfulSwitchAction"
-                    text="colorful"
-                    description="colorful"/>
-        </group>
-
-    </actions>
-
-    <application-components>
-        <component>
-            <implementation-class>com.jiyuanime.ActivatePowerModeApplicationPlugin
-            </implementation-class>
-        </component>
-    </application-components>
-</idea-plugin>
