+++ settings.gradle.kts
@@ -1,0 +1,5 @@
+rootProject.name = "activate-power-mode"
+
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version "1.0.0"
+}
