+++ src/main/java/com/jiyuanime/i18n/ActivatePowerModeBundle.java
@@ -1,0 +1,30 @@
+package com.jiyuanime.i18n;
+
+import com.intellij.DynamicBundle;
+import org.jetbrains.annotations.NonNls;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.PropertyKey;
+
+import java.util.function.Supplier;
+
+public class ActivatePowerModeBundle extends DynamicBundle {
+
+    @NonNls
+    private static final String BUNDLE = "messages.ActivatePowerModeBundle";
+
+    private static final ActivatePowerModeBundle INSTANCE = new ActivatePowerModeBundle();
+
+    private ActivatePowerModeBundle() {
+        super(BUNDLE);
+    }
+
+    @NotNull
+    public static String message(@NotNull @PropertyKey(resourceBundle = BUNDLE) String key, Object @NotNull ... params) {
+        return INSTANCE.getMessage(key, params);
+    }
+
+    @NotNull
+    public static Supplier<String> messagePointer(@NotNull @PropertyKey(resourceBundle = BUNDLE) String key, Object @NotNull ... params) {
+        return INSTANCE.getLazyMessage(key, params);
+    }
+}
