+++ src/main/resources/messages/ActivatePowerModeBundle.properties
@@ -1,0 +1,22 @@
+# Setting
+settings.title=activate-power-mode
+settings.particle.max.count=Particle Max Count
+settings.particle.color=Particle Color
+settings.particle.color.auto=Auto
+settings.shake=Shake
+validation.particle.max.count.greater=The 'particle max count' field must be greater than 0
+validation.particle.max.count.format=The 'particle max count' field format error
+validation.particle.color.required='particle color' is not chosen
+# Action
+action.ActivatePowerModeGroup.text=activate-power-mode
+action.ActivatePowerEnable.text=Enable
+action.ActivatePowerEnable.description=Enable activate power mode
+action.ShakeSwitch.text=Shake
+action.ShakeSwitch.description=Toggle shake effect
+action.ParticleSwitch.text=Particle
+action.ParticleSwitch.description=Toggle particle effect
+action.ComboSwitchAction.text=Combo
+action.ComboSwitchAction.description=Toggle combo display
+action.ColorfulSwitch.text=Colorful
+action.ColorfulSwitch.description=Toggle colorful effect
+
