+++ src/main/resources/messages/ActivatePowerModeBundle_zh.properties
@@ -1,0 +1,21 @@
+# Setting
+settings.title=activate-power-mode
+settings.particle.max.count=\u7C92\u5B50\u6700\u5927\u6570\u91CF
+settings.particle.color=\u7C92\u5B50\u989C\u8272
+settings.particle.color.auto=\u81EA\u52A8
+settings.shake=\u9707\u52A8
+validation.particle.max.count.greater=\u7C92\u5B50\u6700\u5927\u6570\u91CF\u5FC5\u987B\u5927\u4E8E 0
+validation.particle.max.count.format=\u7C92\u5B50\u6700\u5927\u6570\u91CF\u683C\u5F0F\u9519\u8BEF
+validation.particle.color.required=\u672A\u9009\u62E9\u7C92\u5B50\u989C\u8272
+# Action
+action.ActivatePowerModeGroup.text=activate-power-mode
+action.ActivatePowerEnable.text=\u542F\u7528
+action.ActivatePowerEnable.description=\u542F\u7528\u6FC0\u6D3B\u80FD\u91CF\u6A21\u5F0F
+action.ShakeSwitch.text=\u9707\u52A8
+action.ShakeSwitch.description=\u5207\u6362\u9707\u52A8\u6548\u679C
+action.ParticleSwitch.text=\u7C92\u5B50
+action.ParticleSwitch.description=\u5207\u6362\u7C92\u5B50\u6548\u679C
+action.ComboSwitchAction.text=\u8FDE\u51FB
+action.ComboSwitchAction.description=\u5207\u6362\u8FDE\u51FB\u663E\u793A
+action.ColorfulSwitch.text=\u591A\u5F69
+action.ColorfulSwitch.description=\u5207\u6362\u591A\u5F69\u6548\u679C
